resume_parser/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
spacy==3.7.2
python-docx==1.1.0
pdfplumber==0.10.3
pypdf2==3.0.1
pydantic==2.5.0